+++ src/components/certification/index.ts
@@ -3,3 +3,5 @@ export {
   FacilityCertifierDialog,
   UnlinkConfirmDialog,
 } from "./facility-certifier-dialog";
+export { CertifyPanel } from "./certify-panel";
+export { BlueprintList } from "./blueprint-list";